fastapi_graphql_mongo/requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv
pymongo